Move cmd/pivit/import.go -> pkg/pivit/import.go
