+++ pkg/pivit/import.go
@@ -1,4 +1,4 @@
-package main
+package pivit
 
 import (
 	"crypto/x509"
@@ -11,8 +11,8 @@ import (
 	"github.com/pkg/errors"
 )
 
-// commandImport stores a certificate file in a yubikey PIV slot
-func commandImport(file string, first bool, slot string) error {
+// CommandImport stores a certificate file in a yubikey PIV slot
+func CommandImport(file string, first bool, slot string) error {
 	certBytes, err := os.ReadFile(file)
 	if err != nil {
 		return errors.Wrap(err, "read certificate file")
